update docker run examples to 3.0.3
With the separation of ENTRYPOINT and CMD many of the examples fail on "ERROR: Error during SonarQube Scanner execution ERROR: Tasks support was removed in SonarQube 7.6." This is due to the default ENTRYPOINT being sonar_scanner and the example commands then passing sonar_scanner again. Removing this extraneous call (and leaving the -Dsonar.) handles this issue.